pull remote code!

+++ spring-web/src/test/java/org/springframework/http/server/reactive/ServerHttpRequestIntegrationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2019 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -58,7 +58,7 @@ class ServerHttpRequestIntegrationTests extends AbstractHttpHandlerIntegrationTe
 			URI uri = request.getURI();
 			assertThat(uri.getScheme()).isEqualTo("http");
 			assertThat(uri.getHost()).isNotNull();
-			assertThat(uri.getPort()).isNotEqualTo((long) -1);
+			assertThat(uri.getPort()).isNotEqualTo(-1);
 			assertThat(request.getRemoteAddress()).isNotNull();
 			assertThat(uri.getPath()).isEqualTo("/foo");
 			assertThat(uri.getQuery()).isEqualTo("param=bar");
